:construction: Split dataset download by integration or file.
