+++ api/src/services/datasets/download/index.ts
@@ -1,1 +1,3 @@
 export { CreateService } from "./create-service"
+export { CreateFromFileService } from "./create-from-file-service"
+export { CreateFromIntegrationService } from "./create-from-integration-service"
